+++ CHANGELOG.md
@@ -2,8 +2,7 @@ Changelog
 =========
 
 # Unreleased (0.21.0)
-* Fix gRPC server stream handling to be compatible with Java gRPC server by closing
-  the send stream post dispatching the client request.
+* Fix gRPC server stream handling to be compatible with Java gRPC server.
 
 # 0.20.0 (2021-05-18)
 * Add `stream-delay-close-send` option which delays client send stream closure.